Put actually button demo test from input text to button demo
